fix(new post vm): make view model auto dispose
Fixes bugs with showing loading and showing errors when coming back to new post screen.
